Make various bug fixes

+++ src/main/java/com/github/nija123098/evelyn/moderation/temporarychannels/TemporaryChannelCommand.java
@@ -6,16 +6,18 @@ import com.github.nija123098.evelyn.command.ModuleLevel;
 import com.github.nija123098.evelyn.command.annotations.Argument;
 import com.github.nija123098.evelyn.command.annotations.Command;
 import com.github.nija123098.evelyn.config.ConfigHandler;
-import com.github.nija123098.evelyn.discordobjects.wrappers.Channel;
-import com.github.nija123098.evelyn.discordobjects.wrappers.Guild;
-import com.github.nija123098.evelyn.discordobjects.wrappers.VoiceChannel;
+import com.github.nija123098.evelyn.discordobjects.wrappers.*;
 import com.github.nija123098.evelyn.discordobjects.wrappers.event.EventListener;
 import com.github.nija123098.evelyn.discordobjects.wrappers.event.events.DiscordMessageReceived;
+import com.github.nija123098.evelyn.exeption.ArgumentException;
 import com.github.nija123098.evelyn.service.services.ScheduleService;
+import com.github.nija123098.evelyn.util.FormatHelper;
 
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Predicate;
 
 public class TemporaryChannelCommand extends AbstractCommand {
     private static final int TEXT_INACTIVITY = 3_600_000, VOICE_INACTIVITY = 1_800_000, CHECKING_INTERVAL = 60_000, ITERATIONS = VOICE_INACTIVITY / CHECKING_INTERVAL;
@@ -28,7 +30,7 @@ public class TemporaryChannelCommand extends AbstractCommand {
         channels.removeAll(INACTIVITY_MAP.keySet());
         channels.forEach(TemporaryChannelCommand::updateTask);
         ScheduleService.scheduleRepeat(600_000, 600_000, () -> INACTIVITY_MAP.forEach((channel, integer) -> INACTIVITY_MAP.compute(channel, (chan, count) -> {
-            count = !chan.getConnectedUsers().isEmpty() ? 0 : ++count;
+            if (!chan.getConnectedUsers().isEmpty() || chan.getGuild().getUsers().stream().map(User::getPresence).filter(presence -> presence.getStatus() == Presence.Status.ONLINE).map(Presence::getOptionalPlayingText).filter(Optional::isPresent).filter(s -> FormatHelper.filtering(s.get().toLowerCase().replace(" ", "-"), Character::isLetterOrDigit).equals(chan.getName())).count() > ConfigHandler.getSetting(TemporaryGameChannelsConfig.class, chan.getGuild())) return 0;
             if (count > ITERATIONS) {
                 INACTIVITY_MAP.remove(chan);
                 chan.delete();
@@ -39,6 +41,8 @@ public class TemporaryChannelCommand extends AbstractCommand {
     @Command
     public static void command(@Argument(optional = true, replacement = ContextType.NONE, info = "if it's a text channel") Boolean textChannel, @Argument(info = "name") String name, Guild guild){
         Channel channel;
+        if (name.isEmpty()) throw new ArgumentException("Please specify a name for the channel");
+        if (textChannel == null) textChannel = false;
         if (textChannel) {
             channel = guild.createChannel(name);
             updateTask(channel);
